+++ lib/bmg.rb
@@ -40,6 +40,7 @@ module Bmg
   require_relative 'bmg/relation/empty'
   require_relative 'bmg/relation/in_memory'
   require_relative 'bmg/relation/spied'
+  require_relative 'bmg/relation/materialized'
 
   # Deprecated
   Leaf = Relation::InMemory